desktop/windows/requirements.txt
pyserial>=3.5
pyautogui>=0.9.53